Mining Calculator Update v.1.8.1 Visial Mining Calcualtor update. Also a few bugs have been fixed.

+++ emcchain.sql
@@ -52,7 +52,7 @@ CREATE TABLE IF NOT EXISTS `blocks` (
   `numvout` int(10) unsigned DEFAULT NULL,
   `valuein` double DEFAULT NULL,
   `valueout` double DEFAULT NULL,
-  `fee` float DEFAULT NULL,
+  `fee` double DEFAULT NULL,
   `total_coins` double DEFAULT NULL,
   `coindaysdestroyed` double DEFAULT NULL,
   `avgcoindaysdestroyed` double DEFAULT NULL,
@@ -88,7 +88,7 @@ CREATE TABLE IF NOT EXISTS `transactions` (
   `numvout` int(10) unsigned DEFAULT NULL,
   `valuein` double DEFAULT NULL,
   `valueout` double DEFAULT NULL,
-  `fee` float DEFAULT NULL,
+  `fee` double DEFAULT NULL,
   `coindaysdestroyed` double DEFAULT NULL,
   `avgcoindaysdestroyed` double DEFAULT NULL,
   PRIMARY KEY (`id`)
